T1951 ParameterValueSet creation task is completed.
Summary: T1951 ParameterValueSet creation task is completed.

Test Plan:
Review the code. There are several problems that need to be discussed. ParameterValueSet is not created when
PossibleFiniteState is added or PossibleFiniteStateList is added. When PossibleFiniteState is added the server returns
the Parameter with empty "valueSet" property even though it should have at least one value from previous operation.
When PossibleFiniteStateList is added the Parameter and a new ParameterValueSet are not returned at all.

Reviewers: merlin.bieze

Reviewed By: merlin.bieze

Subscribers: sam.gerene

Maniphest Tasks: T1951

Differential Revision: http://phabricator.adgahq.ca/D1168